console: Refactor gtw status panel
