+++ pkg/webui/console/containers/gateway-status-panel/gateway-status-panel.styl
@@ -17,15 +17,6 @@
     background-color: var(--c-bg-error-light)
     color: var(--c-text-error-normal)
 
-  &-unlock-uptime
-    &-message
-      font-weight: $fw.bold
-      margin-top: - $cs.l
-
-    &-img
-      margin-top: $cs.xs
-      margin-left: - $cs.s
-
   &-tooltip
     width: 1.25rem
     height: 1.25rem
